+++ BTCPayServer/Views/Stores/PayButton.cshtml
@@ -1,6 +1,7 @@
 ﻿@model PayButtonViewModel
 @{
     ViewData["Title"] = "Pay Button";
+    ViewData.SetActivePageAndTitle(StoreNavPages.PayButton, "Pay Button");
 }
 
 <section>
